requirements-test.txt
pytest>=7.4.0
pytest-postgresql>=5.0.0
psycopg2-binary>=2.9.0
configobj>=5.0.8
requests>=2.31.0
